delete works but needs refresh
