+++ client/store/itineraries/index.js
@@ -2,7 +2,10 @@ import axios from "axios";
 
 //action types
 const GET_ITINERARIES = "GET_ITINERARIES";
-
+const DELETE_ITINERARY = "DELETE_ITINERARY";
+export const destroyItinerary = (itinerary) => {
+  return { type: DELETE_ITINERARY, itinerary };
+};
 //action creators
 export const getItineraries = (itineraries) => {
   return { type: GET_ITINERARIES, itineraries };
@@ -19,6 +22,16 @@ export const fetchItineraries = (userId) =>
     }
   };
 
+export const deleteItinerary = (itineraryId) =>
+  async function (dispatch) {
+    try {
+      const { data } = await axios.delete(`/api/itinerary/${itineraryId}`);
+      dispatch(destroyItinerary(data));
+    } catch (err) {
+      return err;
+    }
+  };
+
 //reducer
 let initialState = [];
 
@@ -27,7 +40,13 @@ let initialState = [];
 export default function (state = initialState, action) {
   switch (action.type) {
     case GET_ITINERARIES:
-      return action.itineraries
+      return action.itineraries;
+    case DELETE_ITINERARY: {
+      let newState = state.filter(
+        (itinerary) => itinerary.id !== action.itinerary.id
+      );
+      return newState;
+    }
     default:
       return state;
   }
